string and list operators: *, +; real => double; float primitive type
git-svn-id: https://svn.codehaus.org/boo/trunk@217 2c1201b4-01cd-e047-a400-b836ae1fbc61

+++ src/Boo.Antlr/BooExpressionLexer.cs
@@ -34,7 +34,7 @@ namespace Boo.Antlr
 		public const int EOF = 1;
 		public const int NULL_TREE_LOOKAHEAD = 3;
 		public const int TIMESPAN = 4;
-		public const int REAL = 5;
+		public const int DOUBLE = 5;
 		public const int LONG = 6;
 		public const int ESEPARATOR = 7;
 		public const int INDENT = 8;
@@ -589,7 +589,7 @@ _loop6_breakloop:		;
 _loop13_breakloop:								;
 								}    // ( ... )+
 							}
-							_ttype = REAL;
+							_ttype = DOUBLE;
 						}
 						else {
 						}
@@ -674,7 +674,7 @@ _loop13_breakloop:								;
 				}
 _loop19_breakloop:				;
 				}    // ( ... )+
-				_ttype = REAL;
+				_ttype = DOUBLE;
 			}
 			else {
 			}
