+++ src/Boo.Antlr/BooExpressionLexerTokenTypes.cs
@@ -7,7 +7,7 @@ namespace Boo.Antlr
 		public const int EOF = 1;
 		public const int NULL_TREE_LOOKAHEAD = 3;
 		public const int TIMESPAN = 4;
-		public const int REAL = 5;
+		public const int DOUBLE = 5;
 		public const int LONG = 6;
 		public const int ESEPARATOR = 7;
 		public const int INDENT = 8;
